+++ charmpy/__init__.py
@@ -12,7 +12,7 @@ CkNumPes = charm.numPes
 CkExit = charm.exit
 CkAbort = charm.abort
 
-from .entrymethod import when, threaded
+from .entry_method import when, threaded
 
 from .chare import Chare, Group, Array, ArrayMap
